Refactor: Change variable names and add arrayOfClassJobs type to global.d.ts
