+++ src/pages/index.tsx
@@ -2,11 +2,8 @@ import React, { useState, useEffect } from "react";
 import Head from "next/head";
 import { useAuth } from "@/context/AuthContext";
 import { auth } from "@/firebase/firebase";
-import { parseCookies } from "nookies";
-import axios from "axios";
-import { User } from "@/pages/api/api-client";
 
-export default function Home({ initialResults }: any) {
+export default function Home() {
   const { user } = useAuth();
   const [token, setToken] = useState<string | undefined | null>(null);
 
@@ -20,10 +17,6 @@ export default function Home({ initialResults }: any) {
     getToken();
   }, [user, token]);
 
-  useEffect(() => {
-    console.log(initialResults);
-  }, [initialResults]);
-
   return (
     <div>
       <Head>
@@ -31,24 +24,4 @@ export default function Home({ initialResults }: any) {
       </Head>
     </div>
   );
-}
-
-export async function getServerSideProps(context: any) {
-  const { uid, token } = parseCookies(context);
-  const clientRes = await User.getUserInfo(uid, token);
-  let arrayOfClassJobs = null;
-
-  if (clientRes.lodestone_id !== null) {
-    const handlerRes = await axios.get(
-      `https://xivapi.com/character/${clientRes.lodestone_id}`
-    );
-
-    arrayOfClassJobs = handlerRes.data.Character.ClassJobs.slice(0, 19);
-  }
-
-  return {
-    props: {
-      initialResults: arrayOfClassJobs,
-    },
-  };
 }
