+++ types/global.d.ts
@@ -1,5 +1,21 @@
+export type TClassJob = {
+  ClassID: number;
+  ExpLevel: number;
+  ExpLevelMax: number;
+  ExpLevelTogo: number;
+  IsSpecialised: boolean;
+  JobID: number;
+  Level: number;
+  Name: string;
+  UnlockedState: {
+    ID: number;
+    Name: string;
+  };
+};
+
 export type TDashboardProps = {
   lodestoneId: number | null;
+  arrayOfClassJobs: Array<TClassJob>;
 };
 
 export type TRoleAccordionProps = {
